Support fractional leave days and add validation
Updated `CreateAbsentFormHandler` to validate leave days, throwing custom exceptions for invalid or excessive leave. Changed `NumberOfLeave` to `double` in `EmployeeDto` and `AppUser` to support fractional days. Added `AbsentFormException` class for custom exceptions.

+++ EXE201_BE_ThrivoHR.Application/UseCase/V1/Users/EmployeeDto.cs
@@ -24,7 +24,7 @@ public class EmployeeDto : IMapFrom<AppUser>
     public string? Department { get; set; }
     public string? Position { get; set; }
     public DateOnly? DateOfBirth { get; set; }
-    public int NumberOfLeave { get; set; }
+    public double NumberOfLeave { get; set; }
 
     public string? Manager { get; set; }
